requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
duckdb==0.10.3
tabulate==0.9.0
Faker==37.4.2
geopy==2.4.1
airflow-ai-sdk[openai]==0.1.3